refactor (Calendar): simplify and enhance BaseAutocomplete component
