+++ src/app/components/base-components/Autocomplete/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Autocomplete';
